+++ app/models/concerns/token_authority/resourceable.rb
@@ -90,7 +90,7 @@ module TokenAuthority
     def allowed_resources?
       return true unless TokenAuthority.config.rfc_8707_enabled?
 
-      resources.all? { |uri| TokenAuthority.config.rfc_8707_resources.key?(uri) }
+      resources.all? { |uri| TokenAuthority.config.resource_registry.key?(uri) }
     end
 
     # Checks if the current resources are a subset of the granted resources.