+++ spec/requests/token_authority/authorization_grants_controller_spec.rb
@@ -112,7 +112,8 @@ RSpec.describe TokenAuthority::AuthorizationGrantsController, type: :request do
       end
 
       before do
-        allow(TokenAuthority.config).to receive(:rfc_8707_resources).and_return(configured_resources)
+        allow(TokenAuthority.config).to receive(:resource_registry).and_return(configured_resources)
+        allow(TokenAuthority.config).to receive(:rfc_8707_enabled?).and_return(true)
         start_authorization_flow(token_authority_client, resources:)
       end